Revive the project
We update the project in order to catch-up with one year of FreeSpec
development.  In particular, we start using `coq-ext-lib' instead of
`coq-prelude'.  We also use `coqffi` to generate the interface, and
various utility functions.

As of now, the plugin demo is removed, since there are now missing
pieces upstream (including support for [string] conversion).